Add support of 4.09+ multi lines errors in problem matcher
Since ocaml 4.09, when an error covers multiple lines, the message
emitted by the compiler looks like this:

File "robustmatch.ml", lines 33-37, characters 6-23:
Warning 8: this pattern-matching is not exhaustive.
Here is an example of a case that is not matched:
(AB, MAB, A)

This commit detects the presence of an eventual end line. Support of
previous version of the compiler or single line errors is preserved.