Ast exploration for ppx devs
Exploring the OCaml Parsetree

* Live visualization of the OCaml AST inside a separate editor tab.
* The means of exploring the AST both ways: from standard text editor,
  choosing a piece of code and revealing the corresponding AST node
  inside the custom UI; Highlighting the corresponding source code when
  mouse cursor is moved onto a node inside the UI.

This is done by using Custom Editor API that allows defining an
arbitrary output for a certain type of files. Under the hood it uses the
Webview, that can render (almost) any custom HTML, and it communicates
with the extension via message passing.

For the UI itself, i adapted the popular open source option -
astexplorer.net. Notably, besides other minor modifications (redefining
the communications, adding a separate tab for preprocessed AST), I
reduced the application to only contain the Tree view, which is the AST
output part.

Working with preprocessed code and its AST.

* Preview of the source code after it has been transformed by a
  preprocessor.
* A separate mode to navigate the transformed AST (keeping track of the
  original code). Since modern ppxes are (and should be) developed using
  ppxlib, this PR introduces it as a dependency. I use ppxlib for almost
  all AST operations, and most importantly for its version of AST,
  because it's the one that will also be used by ppx developers.